orphans: use Kosaraju's algorithm
This emits exactly the same output as before, but uses
a very different algorithm to find it.

To handle dependency cycles we conceptually collapse all
strongly connected components (generalized cycles) of the
dependency graph to create a condensed graph without any
cycles (a DAG). Find the orphans in this graph and list the
packages in each of the orphan components.

Kosaraju's algorithm to find strongly connected components
is simple and asymptotically optimal, and still runs orders
of magnitudes faster than loading the packages and
dependencies from hawkey.

I've also tried Tarjan's and another algorithm, but couldn't
make them run as fast as Kosaraju's for the relatively small
dependency graphs.

Signed-off-by: Emil Renner Berthing <esmil@mailme.dk>