Add repoclosure --check option
This option, like the --pkg option, narrows the scope of the dependency
closure check, but at the repository level rather than the package level.
This is particularly useful for add-on repositories, where you want to
check for dependency issues within the add-on repository itself but are
not concerned about issues within the underlying base repository, which
you may have no control over to fix things anyway. Limiting checks to
the repository or repositories of interest makes it much faster to check
than doing a full check over all repositories too.