Use latest package across all repos
Using filter() with latest=True only appears to limit packages to
the latest in each repo, so if packages of the same name appear in
multiple repos, the latest from each is included in the result. This
can happen, for instance, with a base and updates repository set.
Multilib seems to confuse the situation further on occasion. Whether
this is intentional or a bug, I don't know, but this local code,
borrowed from dnf itself, produces the desired result of filtering
down to the latest package of each name across all repos.