Do not abort if tor cannot be contacted, just log a warning
Since tor not working mean that the socks proxy would fail,
we do not need to fail ourself.

However, one case where this would cause trouble is if someone
replace the tor process by something else, but I will
assume that anonimity would be compromissed already since a long
time if that happen.

It was found that this prevent dnf system-upgrade from working, since
tor is not running when doing the package upgrade after the reboot.