tracer: drop TracerCommand implementation
Tracer parameters may now be in conflict with DNF parameters.
Same parameter names with relevance on their order would be
unnecessarily confusing for the user and renaming tracer arguments
is not a solution, but a temporary workaround, because DNF can add
new conflicting arguments in the future. Therefore I would suggest
to use standalone tracer instead of TracerCommand