dnf-system-upgrade.service: make sure systemd waits for us to finish
In f0bef8e9193881f388 I forgot to add the ordering dependency
with system-update.target. This was "wrong", but did not matter until
system-update-cleanup.service which is started after system-update.target
was added in systemd-233. Fix this omission.

(This is the same bug as
https://bugzilla.redhat.com/show_bug.cgi?id=1430920. It wasn't reported
so far, but dnf-system-upgrade.service is subject to the same race as
pk-offline-update.service.)

Also: add " using DNF" to Description, so it's clearer which upgrade
mechanism is started when multiple are installed.

Closes: #89
Approved by: ignatenkobrain