Remove --datadir option from system upgrade
The former behavior of system-upgrade installs all packages from given
directory, but it looses information about origin of packages. The present
behavior uses directly cached packages in repo, that also allows to use local
repositories. Unfortunately the problems remains with usage of --datadir option
and cannot be fixed, therefore it will be better to not use it at all.
The option --datadir is also redundant with conf option to redirect of dnf
cache.