system-upgrade: record and reapply removals (RhBug:1758588)
As discussed in RHBZ #1758588, the transactions produced at the
`download` and `update` steps can be different in some cases due
to the `DISTUPGRADE` rule being applied to the `download`
transaction but not the `update` transaction. Recording the
removals as well as the installs in the `download` transaction
and re-applying them to the `update` transaction should resolve
at least most of these cases. However, a more correct long-term
fix for this would be for dnf to provide a supported mechanism for
serializing and de-serializing a transaction object, and for the
plugin to use that instead.

Signed-off-by: Adam Williamson <awilliam@redhat.com>