system-upgrade: Use Transaction Store/Replay
Uses the Transaction Store/Replay feature for the system upgrade
transaction. Keeps the original state file, but removes the items that
are no longer necessary from it. The transaction itself is stored
separately in its own json.

Also moves the state file from /var/lib/dnf/system-upgrade.json to
/var/lib/dnf/system-upgrade/system-upgrade-state.json, so that all the
files are in one place (this will leave a stale
/var/lib/dnf/system-upgrade.json present on systems where this was
present before this change).