system-upgrade: Use installroot for data dir
Prepends installroot to the system-upgrade's data dir path (where dnf
cache and the plugin's state/transaction files are stored). This
doesn't have much of a practical use, since the plugin workflow involves
a reboot of the system being upgraded (and upgrading an installroot can
be done directly without the need for this plugin). But conceptually the
data belong inside the installroot and it is useful for tests.