test_system_upgrade: Set installroot in the mocked cli
Fixes running tests on Fedora 31.

Sets the installroot to the default "/" in the mocked cli.base.conf. The
whole object is mocked and on Fedora 31 the mocked installroot object wasn't
resulting in anything the command could use. Since all the directories
are already changed to temporary locations, we don't need to set the
installroot to anything special.