Power off when system upgrade is done
Hi there,

I suspect that most users will probably run a system upgrade at a time when they know that they won't be using the machine for some time. This is because they don't know in advance how long a system upgrade is going to take. This is probably at the end of the work day or before they go to bed, so in all likelihood the machine is going to sit there the whole night and consume power for no reason. I think it would be better to instead just power off the machine. Do you agree with my reasoning?