Cleaning up destdir is not necessary, dnf does it (RhBug:2024430)
dnf automatically cleans up packages used in a transaction, this can be
configured by `keepcache` option but we have to preserve the persistor
which contains their paths.

The persistor file remains after each run of dnf so it remains either
way, if we delete it dnf creates it again when finalizing base.

cachedir still gets cleaned up but it is hardcoded to:
`var/lib/dnf/system-upgrade` and cannot be overwritten.

For: https://github.com/rpm-software-management/dnf-plugins-extras/issues/200
https://bugzilla.redhat.com/show_bug.cgi?id=2024430